use PyQtAxis
