+++ whitecanvas/backend/pyqtgraph/canvas.py
@@ -11,7 +11,7 @@ from qtpy import QtCore, QtGui
 from qtpy.QtCore import Signal
 
 from whitecanvas import protocols
-from whitecanvas.backend.pyqtgraph._base import InsetPlotItem, PyQtLayer
+from whitecanvas.backend.pyqtgraph._base import InsetPlotItem, PyQtAxis, PyQtLayer
 from whitecanvas.backend.pyqtgraph._labels import Axis, AxisLabel, Ticks, Title
 from whitecanvas.backend.pyqtgraph._legend import QtItemSampleBase, make_sample_item
 from whitecanvas.backend.pyqtgraph._qt_utils import from_qt_button, from_qt_modifiers
@@ -112,14 +112,17 @@ class Canvas:
 
     def _plt_twinx(self) -> Canvas:
         """Create a twinx canvas"""
-        plotitem = self._plot_item
-        vb1 = plotitem.vb
+        vb1 = self._plot_item.vb
         vb2 = pg.ViewBox()
-        canvas = Canvas(pg.PlotItem(viewBox=vb2), yaxis="right")
+        new_item = pg.PlotItem(
+            viewBox=vb2,
+            axisItems={"right": PyQtAxis("right"), "bottom": PyQtAxis("bottom")},
+        )
+        canvas = Canvas(new_item, yaxis="right")
 
         self._get_scene().addItem(vb2)
-        plotitem.getAxis("right").linkToView(vb2)
-        vb2.setXLink(plotitem)
+        self._plot_item.getAxis("right").linkToView(vb2)
+        vb2.setXLink(self._plot_item)
 
         def _update_views():
             vb2.setGeometry(vb1.sceneBoundingRect())
@@ -131,14 +134,16 @@ class Canvas:
         return canvas
 
     def _plt_twiny(self) -> Canvas:
-        plotitem = self._plot_item
-        vb1 = plotitem.vb
+        vb1 = self._plot_item.vb
         vb2 = pg.ViewBox()
-        canvas = Canvas(pg.PlotItem(viewBox=vb2), xaxis="top")
+        new_item = pg.PlotItem(
+            viewBox=vb2, axisItems={"left": PyQtAxis("left"), "top": PyQtAxis("top")}
+        )
+        canvas = Canvas(new_item, xaxis="top")
 
         self._get_scene().addItem(vb2)
-        plotitem.getAxis("bottom").linkToView(vb2)
-        vb2.setYLink(plotitem)
+        self._plot_item.getAxis("bottom").linkToView(vb2)
+        vb2.setYLink(self._plot_item)
 
         def _update_views():
             vb2.setGeometry(vb1.sceneBoundingRect())
@@ -151,7 +156,10 @@ class Canvas:
 
     def _plt_inset(self, rect: Rect) -> Canvas:
         vb = pg.ViewBox()
-        item = pg.PlotItem(viewBox=vb)
+        item = pg.PlotItem(
+            viewBox=vb,
+            axisItems={"left": PyQtAxis("left"), "bottom": PyQtAxis("bottom")},
+        )
         canvas = Canvas(item)
         inset_item = InsetPlotItem(item, rect)
         inset_item.setParentItem(self._plot_item.vb)
